Remove proposals from db which removed from chain
